refactoring - separate Factory class out
